Created StorageRetrievalAsyncTask class, and fixed bug in detail screen for retrieving/deleting current favorite article.
More specifically this bug was introduced when we made the StorageIntentService class.  The actual bug was that checking only by guid, or deleting only by guid was no longer sufficient, as we now needed to check the type of article, since now there are not only favorite type articles, but category type too on the same table.  Thus we needed to take into account the type as well, and not only the guid (composite key guid+type, insted of just guid before)